refactor: merge LoginForm into LoginPage and fix selector sorting compatibility
