--- src/components/login-form/login-form.css
@@ -1,11 +1,0 @@
-.login__error-message {
-  color: #ff4400;
-  margin-bottom: 10px;
-  text-align: center;
-  font-size: 14px;
-}
-
-.login__error-message span {
-  display: block;
-  margin-bottom: 4px;
-}
