+++ src/store/reviews/reviews-selectors.ts
@@ -19,7 +19,8 @@ export const selectSendingError = (state: State): string | null =>
 
 export const selectSortedReviews = createSelector(
   [selectReviews],
-  (reviews) => reviews
-    .toSorted((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime())
-    .slice(0, MAX_REVIEWS)
+  (reviews) =>
+    [...reviews]
+      .sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime())
+      .slice(0, MAX_REVIEWS)
 );
